done with the job Seeker Controller

+++ src/main/java/com/mananluvtocode/jobportal/repository/JobSeekerProfileRepository.java
@@ -3,6 +3,6 @@ package com.mananluvtocode.jobportal.repository;
 import com.mananluvtocode.jobportal.entity.JobSeekerProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobSeekerProfileRepository extends JpaRepository<JobSeekerProfile, Long> {
+public interface JobSeekerProfileRepository extends JpaRepository<JobSeekerProfile, Integer> {
 
 }
